Update stager.sh
more code cleanup

+++ stager.sh
@@ -56,7 +56,7 @@ if [ -n "$RPIVER" ] && [ "$RPIVER" -lt 5 ]; then
 
    apt-get install docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin -y
    systemctl restart docker
-elif [ "$RPIVER" -eq 5 ]; then
+elif [ -n "$RPIVER" ] && [ "$RPIVER" -eq 5 ]; then
    for pkg in docker.io docker-doc docker-compose docker-compose-v2 podman-docker containerd runc; do sudo apt-get remove $pkg; done
    # Add Docker's official GPG key:
    apt-get update
